move test to fsjournal
