+++ journal/fsjournal/fs_test.go
@@ -1,6 +1,7 @@
-package journal
+package fsjournal
 
 import (
+	"container/list"
 	"os"
 	"path/filepath"
 	"testing"
@@ -29,4 +30,18 @@ func TestRollingRemovesOldFiles(t *testing.T) {
 	// so we should still have only the maximum kept files.
 	files, _ := os.ReadDir(dir)
 	req.Lenf(files, j.keep, "files are not being pruned from the journal directory")
+}
+
+func newtestfsjournal(t *testing.T, lr repo.LockedRepo, sizeLimit int64, keep int) *fsJournal {
+	req := require.New(t)
+	dir := filepath.Join(lr.Path(), "journal")
+	req.NoErrorf(os.MkdirAll(dir, 0755), "could not make journal directory")
+
+	j := &fsJournal{
+		dir:       dir,
+		sizeLimit: sizeLimit,
+		keep:      keep,
+		old:       list.New(),
+	}
+	return j
 }
